feat: Adição da tela de usuários desativados e ajuste nos arquivos
