+++ FrontEnd/components/usuarios/modalUsuario/modalRedefinirSenha.vue
@@ -8,7 +8,7 @@
               Redefinir Senha
             </v-card-text>
             <v-btn text icon @click="$emit('fechaModal')">
-              <img src="../../assets/img/xmark.png" alt="close" />
+              <img src="~/assets/img/xmark.png" alt="close" />
             </v-btn>
           </v-toolbar>
         </div>
@@ -67,7 +67,7 @@
 import {
   senhaRules,
   confirmarSenhaRule,
-} from "../../services/validationsRules";
+} from "~/services/validationsRules";
 
 export default {
   props: ["usuario"],
